rename that and add in other output related calls
